api/requirements.txt
djangorestframework
django-cors-headers
django-phonenumber-field
gunicorn
psycopg2
pyrebase4
python-dotenv
yapf
